refactor(desktop): extract pane manager hook
